backend/requirements.txt
fastapi~=0.67.0
pydantic~=1.8.2
uvicorn~=0.14.0
pandas~=1.3.1
sklearn~=0.0
python-igraph~=0.9.6
loguru~=0.5.3
